Address review: extract ExpiryInput, label via FormGroup, PatternFly Flex layout
