+++ src/Slices/Settings/UI/Tabs/Token/TokenForm.tsx
@@ -5,8 +5,6 @@ import {
   ExpandableSection,
   Flex,
   FlexItem,
-  FormSelect,
-  FormSelectOption,
   HelperText,
   HelperTextItem,
   InputGroup,
@@ -20,107 +18,7 @@ import styled from "styled-components";
 import { ClientType } from "@/Core";
 import { AppAlert, ClipboardCopyButton, Description } from "@/UI/Components";
 import { words } from "@/UI/words";
-
-// Preset token lifetimes in seconds; the empty default means no explicit expiry.
-const EXPIRY_OPTIONS: { seconds: number; label: string }[] = [
-  { seconds: 3600, label: "1 hour" },
-  { seconds: 86400, label: "1 day" },
-  { seconds: 604800, label: "7 days" },
-  { seconds: 2592000, label: "30 days" },
-  { seconds: 31536000, label: "1 year" },
-];
-
-const CUSTOM_UNITS: { unit: string; seconds: number }[] = [
-  { unit: "minutes", seconds: 60 },
-  { unit: "hours", seconds: 3600 },
-  { unit: "days", seconds: 86400 },
-];
-
-interface ExpiryInputProps {
-  onChange(value: number | null): void;
-  isDisabled: boolean;
-}
-
-/**
- * Expiry editor for the create-token form: preset lifetimes plus a custom amount + unit entry.
- * Emits the effective lifetime in seconds through onChange, or null when no (valid) expiry is set.
- */
-const ExpiryInput: React.FC<ExpiryInputProps> = ({ onChange, isDisabled }) => {
-  const [choice, setChoice] = useState("");
-  const [customAmount, setCustomAmount] = useState("");
-  const [customUnit, setCustomUnit] = useState("days");
-
-  const toSeconds = (choice: string, amount: string, unit: string): number | null => {
-    if (choice === "") {
-      return null;
-    }
-
-    if (choice !== "custom") {
-      return Number(choice);
-    }
-
-    const parsed = Number(amount);
-    const unitSeconds = CUSTOM_UNITS.find((option) => option.unit === unit)?.seconds;
-
-    return Number.isInteger(parsed) && parsed > 0 && unitSeconds ? parsed * unitSeconds : null;
-  };
-
-  const update = (choice: string, amount: string, unit: string) => {
-    setChoice(choice);
-    setCustomAmount(amount);
-    setCustomUnit(unit);
-    onChange(toSeconds(choice, amount, unit));
-  };
-
-  return (
-    <>
-      <FlexItem>
-        <FormSelect
-          id="token-expiry"
-          aria-label="ExpiryOption"
-          value={choice}
-          onChange={(_event, value) => update(value, customAmount, customUnit)}
-          isDisabled={isDisabled}
-        >
-          <FormSelectOption value="" label={words("settings.tabs.token.expiry.never")} />
-          {EXPIRY_OPTIONS.map((option) => (
-            <FormSelectOption
-              key={option.seconds}
-              value={String(option.seconds)}
-              label={option.label}
-            />
-          ))}
-          <FormSelectOption value="custom" label={words("settings.tabs.token.expiry.custom")} />
-        </FormSelect>
-      </FlexItem>
-      {choice === "custom" && (
-        <FlexItem>
-          <InputGroup>
-            <AmountInput
-              id="token-expiry-amount"
-              type="number"
-              aria-label="ExpiryCustomAmount"
-              value={customAmount}
-              onChange={(_event, value) => update(choice, value, customUnit)}
-              isDisabled={isDisabled}
-            />
-            <UnitSelect
-              id="token-expiry-unit"
-              aria-label="ExpiryCustomUnit"
-              value={customUnit}
-              onChange={(_event, value) => update(choice, customAmount, value)}
-              isDisabled={isDisabled}
-            >
-              {CUSTOM_UNITS.map((option) => (
-                <FormSelectOption key={option.unit} value={option.unit} label={option.unit} />
-              ))}
-            </UnitSelect>
-          </InputGroup>
-        </FlexItem>
-      )}
-    </>
-  );
-};
+import { ExpiryInput } from "./ExpiryInput";
 
 interface AdvancedSectionProps {
   toggleText: string;
@@ -148,40 +46,50 @@ const AdvancedSection: React.FC<AdvancedSectionProps> = ({
       isExpanded={isExpanded}
       onToggle={(_event, expanded) => setIsExpanded(expanded)}
     >
-      <AdvancedContent>
-        <Title headingLevel="h3" size="md">
-          {words("settings.tabs.token.clientTypes")}
-        </Title>
-        <ToggleGroup aria-label="ClientTypes">
-          <ToggleGroupItem
-            icon={<KeyIcon />}
-            text="api"
-            aria-label="ApiOption"
-            isSelected={isClientTypeSelected("api")}
-            onChange={(_event, selected) => getClientTypeSelector("api")(selected)}
-            isDisabled={isBusy}
-          />
-          <ToggleGroupItem
-            icon={<RobotIcon />}
-            text="agent"
-            aria-label="AgentOption"
-            isSelected={isClientTypeSelected("agent")}
-            onChange={(_event, selected) => getClientTypeSelector("agent")(selected)}
-            isDisabled={isBusy}
-          />
-          <ToggleGroupItem
-            icon={<CodeIcon />}
-            text="compiler"
-            aria-label="CompilerOption"
-            isSelected={isClientTypeSelected("compiler")}
-            onChange={(_event, selected) => getClientTypeSelector("compiler")(selected)}
-            isDisabled={isBusy}
-          />
-        </ToggleGroup>
-        <HelperText>
-          <HelperTextItem>{words("settings.tabs.token.clientTypes.help")}</HelperTextItem>
-        </HelperText>
-      </AdvancedContent>
+      <Flex
+        direction={{ default: "column" }}
+        alignItems={{ default: "alignItemsFlexStart" }}
+        gap={{ default: "gapSm" }}
+      >
+        <FlexItem>
+          <Title headingLevel="h3" size="md">
+            {words("settings.tabs.token.clientTypes")}
+          </Title>
+        </FlexItem>
+        <FlexItem>
+          <ToggleGroup aria-label="ClientTypes">
+            <ToggleGroupItem
+              icon={<KeyIcon />}
+              text="api"
+              aria-label="ApiOption"
+              isSelected={isClientTypeSelected("api")}
+              onChange={(_event, selected) => getClientTypeSelector("api")(selected)}
+              isDisabled={isBusy}
+            />
+            <ToggleGroupItem
+              icon={<RobotIcon />}
+              text="agent"
+              aria-label="AgentOption"
+              isSelected={isClientTypeSelected("agent")}
+              onChange={(_event, selected) => getClientTypeSelector("agent")(selected)}
+              isDisabled={isBusy}
+            />
+            <ToggleGroupItem
+              icon={<CodeIcon />}
+              text="compiler"
+              aria-label="CompilerOption"
+              isSelected={isClientTypeSelected("compiler")}
+              onChange={(_event, selected) => getClientTypeSelector("compiler")(selected)}
+              isDisabled={isBusy}
+            />
+          </ToggleGroup>
+        </FlexItem>
+        <FlexItem>
+          <HelperText>
+            <HelperTextItem>{words("settings.tabs.token.clientTypes.help")}</HelperTextItem>
+          </HelperText>
+        </FlexItem>
+      </Flex>
     </ExpandableSection>
   );
 };
@@ -213,12 +121,11 @@ export const TokenForm: React.FC<Props> = ({
 }) => (
   <Container>
     <Description>{words("settings.tabs.token.description")}</Description>
-    <PaddedFlex>
+    <PaddedFlex alignItems={{ default: "alignItemsFlexEnd" }}>
       <FlexItem>
-        <label htmlFor="token-expiry">{words("settings.tabs.token.expiry")}</label>
+        {/* An idempotent (non-revocable) token carries no time-based claims, so it cannot expire. */}
+        <ExpiryInput onChange={onExpireChange} isDisabled={isBusy || !isRevocable} />
       </FlexItem>
-      {/* An idempotent (non-revocable) token carries no time-based claims, so it cannot expire. */}
-      <ExpiryInput onChange={onExpireChange} isDisabled={isBusy || !isRevocable} />
       <FlexItem>
         <Checkbox
           id="token-revocable"
@@ -274,21 +181,6 @@ const StyledInputGroup = styled(InputGroup)`
   padding-top: 1rem;
   padding-bottom: 1rem;
   max-width: 600px;
-`;
-
-const AdvancedContent = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 0.5rem;
-`;
-
-const AmountInput = styled(TextInput)`
-  width: 6rem;
-`;
-
-const UnitSelect = styled(FormSelect)`
-  width: auto;
 `;
 
 const Container = styled.div`
